+++ ftplugin/elixir.lua
@@ -1,0 +1,1 @@
+require("primebrook.keymaps.elixir").setup()
