+++ lua/primebrook/keymaps/sql.lua
@@ -1,0 +1,8 @@
+local M = {}
+
+function M.setup()
+	-- formatting
+	vim.keymap.set("n", "<leader>fs", ":<C-U>%!pg_format -<CR>", { noremap = true, silent = true })
+end
+
+return M
